Rename and update references to files
