+++ lib/airtap.js
@@ -6,9 +6,9 @@ var EventEmitter = require('events').EventEmitter
 
 var createControlServer = require('./control-app')
 var setup_test_instance = require('./setup')
-var SauceBrowser = require('./SauceBrowser')
-var PhantomBrowser = require('./PhantomBrowser')
-var Electron = require('./Electron')
+var SauceBrowser = require('./sauce-browserlist')
+var PhantomBrowser = require('./phantom-browserlist')
+var Electron = require('./electron-browserlist')
 
 module.exports = Zuul
