+++ lib/electron-browserlist.js
