+++ lib/flatten-browserlist.js
